Cleanup Branch

See merge request cse1105/2019-2020/repos/oopp-group-39/repository-template!43

+++ server/src/main/java/nl/tudelft/oopp/group39/building/exceptions/BuildingExistsException.java
@@ -1,7 +1,7 @@
 package nl.tudelft.oopp.group39.building.exceptions;
 
 public class BuildingExistsException extends RuntimeException {
-    public BuildingExistsException(int id) {
+    public BuildingExistsException(Long id) {
         super("Building with id " + id + " already exist.");
     }
 }
