+++ server/src/main/java/nl/tudelft/oopp/group39/facility/exceptions/FacilityExistsException.java
@@ -1,7 +1,7 @@
 package nl.tudelft.oopp.group39.facility.exceptions;
 
 public class FacilityExistsException extends RuntimeException {
-    public FacilityExistsException(int id) {
+    public FacilityExistsException(Long id) {
         super("Facility with id " + id + " already exist.");
     }
 }
