+++ server/src/main/java/nl/tudelft/oopp/group39/facility/exceptions/FacilityNotFoundException.java
@@ -1,7 +1,7 @@
 package nl.tudelft.oopp.group39.facility.exceptions;
 
 public class FacilityNotFoundException extends RuntimeException {
-    public FacilityNotFoundException(int id) {
+    public FacilityNotFoundException(Long id) {
         super("Facility with id " + id + " wasn't found.");
     }
